refactor(markdown): share KaTeX HTML processor wiring across webview and export
The progress-view formatter and the HTML chat export each rebuilt the same
markdown-it + KaTeX + hljs pipeline: engine registration, project macro set,
highlight hook, and processor caching, differing only in error colour and the
export's LaTeX-reference formatter.

Add @shared/markdown/katexHtmlProcessor (direct-path import, kept out of the
barrel for the same side-effect reason as texmathPlugin) exposing exactly
those two knobs, and move katexMacros into src/shared/markdown so the export
no longer reaches into the progress-view frontend for it.
